incorrect handling of token decimals in addToken

+++ src/interfaces/ITokenRegistry.sol
@@ -45,7 +45,12 @@ interface ITokenRegistry {
     /// @param oldPrice The old price for the token
     /// @param newPrice The new price for the token
     /// @param updater The address of the user who updated the price
-    event TokenPriceUpdated(IERC20 indexed token, uint256 oldPrice, uint256 newPrice, address indexed updater);
+    event TokenPriceUpdated(
+        IERC20 indexed token,
+        uint256 oldPrice,
+        uint256 newPrice,
+        address indexed updater
+    );
 
     /// @notice Error thrown when an operation is attempted on an unsupported token
     /// @param token The address of the unsupported token
@@ -65,7 +70,11 @@ interface ITokenRegistry {
     /// @param token The address of the token to add
     /// @param decimals The number of decimals for the token
     /// @param initialPrice The initial price for the token
-    function addToken(IERC20 token, uint256 decimals, uint256 initialPrice) external;
+    function addToken(
+        IERC20 token,
+        uint8 decimals,
+        uint256 initialPrice
+    ) external;
 
     /// @notice Removes a token from the registry
     /// @param token The address of the token to remove
@@ -85,13 +94,19 @@ interface ITokenRegistry {
     /// @param token The address of the token to convert
     /// @param amount The amount of tokens to convert
     /// @return The converted amount in the unit of account
-    function convertToUnitOfAccount(IERC20 token, uint256 amount) external view returns (uint256);
+    function convertToUnitOfAccount(
+        IERC20 token,
+        uint256 amount
+    ) external view returns (uint256);
 
     /// @notice Converts an amount in the unit of account to a token amount
     /// @param token The address of the token to convert to
     /// @param amount The amount in the unit of account to convert
     /// @return The converted amount in the specified token
-    function convertFromUnitOfAccount(IERC20 token, uint256 amount) external view returns (uint256);
+    function convertFromUnitOfAccount(
+        IERC20 token,
+        uint256 amount
+    ) external view returns (uint256);
 
     /// @notice Retrieves the list of supported tokens
     /// @return An array of addresses of supported tokens
@@ -100,5 +115,7 @@ interface ITokenRegistry {
     /// @notice Retrieves the information for a specific token
     /// @param token The address of the token to get information for
     /// @return TokenInfo struct containing the token's information
-    function getTokenInfo(IERC20 token) external view returns (TokenInfo memory);
+    function getTokenInfo(
+        IERC20 token
+    ) external view returns (TokenInfo memory);
 }
